+++ Network/layers/FlatteningLayer.cpp
@@ -11,18 +11,18 @@ cn::FlatteningLayer::FlatteningLayer(int _id, Network &_network) : Layer(_id, _n
     outputSize = Vector3<int>(size, 1, 1);
 }
 
-cn::Bitmap<float> cn::FlatteningLayer::run(const cn::Bitmap<float> &input) {
+cn::Bitmap<double> cn::FlatteningLayer::run(const cn::Bitmap<double> &input) {
     if(input.size().multiplyContent() != inputSize.multiplyContent())
         throw std::logic_error("invalid input input for flattening layer!");
-    return Bitmap<float>({inputSize.multiplyContent(), 1, 1}, input.data());
+    return Bitmap<double>({inputSize.multiplyContent(), 1, 1}, input.data());
 }
 
-float cn::FlatteningLayer::getChain(const Vector3<int> &inputPos) {
+double cn::FlatteningLayer::getChain(const Vector3<int> &inputPos) {
     if(getMemoState(inputPos)){
         return getMemo(inputPos);
     }
     int outputIndex = network->getInput(__id)->getDataIndex(inputPos);
-    float res = network->getChain(__id + 1, {outputIndex, 0, 0});
+    double res = network->getChain(__id + 1, {outputIndex, 0, 0});
     setMemo(inputPos, res);
     return res;
 }
